PrintF concluida

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: bchagas- <bchagas-@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/12 05:30:34 by bchagas-          #+#    #+#             */
-/*   Updated: 2025/08/17 21:49:36 by bchagas-         ###   ########.fr       */
+/*   Updated: 2025/08/21 07:18:04 by bchagas-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@
 
 int	ft_putchar(char c);
 int	ft_putstr(char *s);
-int	ft_putnbr(int n);
+int	ft_putnbr(long long n);
 int	ft_puthex(unsigned long n, char upper);
 int	ft_putptr(void *ptr);
 int	ft_printf(const char *format, ...);
